The owner's call (2026-09-18): the Codex checkout's uncommitted layer is
the real MossTank state, its turning, buff, loot and inventory work
included; only the DotRecast pipeline goes. So the fold now stands on that
layer (the snapshot minus DotRecast) with everything from
claude/mosstank-on-main merged in.

Resolution: the contract keeps both sides' additions; Runtime keeps main's
appraisal-origin model (PresentInUi) over the WIP's equivalent, and the
WIP's redundant origin check in the appraisal window goes; the WIP's
surface additions (timed enchantments, busy state, combat control, forget
reported, contents ready, ammo and tinker fields) ride on main's surface;
the graphical surface subclass is gone with main.

Suites: Core, Runtime, Headless green; App 1 red (walk populator
allocation pin); MossTank 52 red, the WIP's own worklist, next.

+++ src/AcDream.Plugins.MossTank/CombatController.cs
@@ -2695,7 +2695,7 @@ internal sealed class CombatController
                     && message.SenderObjectId != 0u
                     && message.SenderObjectId
                         == _host.Automation.Character.ObjectId,
-                logTextType: message.LogTextType);
+                logTextType: (uint)message.LogTextType);
         }
     }
